[chore]add check for columns in users table

+++ db/migrate/20140328040948_add_user_name_to_users.rb
@@ -1,7 +1,15 @@
 class AddUserNameToUsers < ActiveRecord::Migration
   def change
-    add_column :users, :user_name, :string
-    remove_column :users, :name
-    remove_column :users, :subdomain
+    unless column_exists?(:users, :user_name)
+      add_column :users, :user_name, :string
+    end
+
+    if column_exists?(:users, :name)
+      remove_column :users, :name
+    end
+
+    if column_exists? :users, :subdomain
+      remove_column :users, :subdomain
+    end
   end
 end
